feat(sdk): Implement from_connection_string() for QuicClient
Implement "from_connection_string()" API for "QuicClient". The main
difference between the implementation for this and the implementation
for "TcpClient" differs in "parse_options()", where the former consists
of more options and TLS isn't supported yet. Mark it as TODO and disabled
it in "FromStr" for "ConnectionString<QuicConnectionStringOptions>".
Open it once TLS support it implemented for "QuicClient".

Basic usage example is shown in the following:
```
let client = QuicClient::from_connection_string("iggy+quic://iggy:iggy@127.0.0.1:8080")?;
client.connect().await?;
client.login_user("iggy","iggy").await?;
Ok(())
```

It doesn't seem to be able to retrieve client address or server name
from connection string without specifying it in options, however,
specifying these in options would be non-sense as it doesn't fit the
connection string convention, server name and client address shouldn't
be treated as options.

They remain the same as the one used in default(), discussions are needed
to see whether it's needed to implement other helper or enable more
mechanism to get these value from connection string or not.

Signed-off-by: I Hsin Cheng <richard120310@gmail.com>