feat(sdk): Support PAT for ConnectionString
Implement PAT support for ConnectionString. Currently only
"UsernamePassword" is supported for usage of ConnectionString.
Since PAT is already implemented in the server and SDK, support
it in ConnectionString can be a finishing touch.

Determine whether the user-provided credentials length is equal
to 1, if so, ConnectionString assume the user is trying to provide
a PAT and use it as the login method.

When the credentials can be separated in 2 parts with ":", it means
the user is trying to login with UsernamePassword. Otherwise, it's
an invalid connection string.

Signed-off-by: I Hsin Cheng <richard120310@gmail.com>